Hospitalmanagementsystem/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
decorator==5.1.1
Django==5.0.6
django-extensions==3.2.3
django-filter==24.2
djangorestframework==3.15.2
executing==2.0.1
ipdb==0.13.13
ipython==8.25.0
jedi==0.19.1
matplotlib-inline==0.1.7
nepali-datetime==1.0.8.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
wcwidth==0.2.13
